Flask_Application/Flask/requirements_mod.txt
./application/util/PyGeoApiExtended
affine==2.3.1
appdirs==1.4.4
APScheduler==3.7.0
arcgis2geojson==2.0.1
arrow==1.1.1
async-generator==1.10
attrs==21.4.0
Babel==2.9.1
beautifulsoup4==4.11.1
boto3==1.18.21
botocore==1.21.21
Bottleneck==1.3.4
cachetools==4.2.2
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.4
cligj==0.7.2
colorama==0.4.4
cssselect==1.1.0
cycler==0.11.0
Fiona
Flask==2.0.3
Flask-BasicAuth==0.2.0
Flask-Cors==3.0.10
Flask-HTTPAuth==4.4.0
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
fonttools==4.25.0
future-annotations==1.0.0
GeoAlchemy2==0.9.3
geojson==2.5.0
geopandas
google-api-core==1.31.1
google-api-python-client==2.15.0
google-auth==1.34.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
greenlet==1.1.1
h11==0.13.0
httplib2==0.19.1
idna==3.2
importlib-metadata==4.11.3
importlib-resources==5.7.1
itsdangerous==2.1.2
Jinja2==3.0.3
jmespath==0.10.0
joblib==1.1.0
jsonschema==4.4.0
kiwisolver==1.3.2
lxml==4.8.0
mapclassify==2.4.3
MarkupSafe==2.1.1
matplotlib==3.5.1
multidict==5.1.0
munch==2.5.0
munkres==1.1.4
numexpr==2.7.3
oauth2client==4.1.3
outcome==1.1.0
overpass==0.7
packaging==21.3
pandas
pdfminer.six==20200517
pdfplumber==0.5.28
Pillow==9.0.1
pip==21.2.3
protobuf==3.17.3
psycopg2-binary==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.10.1
pydantic==1.9.0
PyDrive==1.3.1
pyee==8.2.2
pygeoapi==0.12.0
pyOpenSSL==22.0.0
pyparsing==2.4.7
pyppeteer==1.0.2
pyquery==1.4.3
pyrsistent==0.18.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.3
PyYAML==6.0
rasterio==1.2.10
requests==2.26.0
rsa==4.7.2
Rtree==0.9.7
s3transfer==0.5.0
selenium==4.1.5
setuptools==62.1.0
Shapely==1.7.1
simplejson==3.17.3
six==1.16.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.22
sqlathanor==0.7.0
stravalib==0.10.2
tokenize-rt==4.2.1
topojson==1.1
typing_extensions==4.2.0
tzlocal==2.1
uritemplate==3.0.1
urllib3==1.26.6
validator-collection==1.5.0
w3lib==1.22.0
Wand==0.6.6
websockets==10.3
Werkzeug==2.1.2
wheel==0.37.0
wincertstore==0.2
wsproto==1.1.0
WTForms==2.3.3
zipp==3.8.0
